backend/requirements.txt
fastapi>=0.110.0
python-dotenv>=1.0.0
uvicorn>=0.29.0
pydantic>=2.9.2
pydantic-settings>=2.5.0
google-adk
google-genai
google-cloud-storage>=2.16.0
google-cloud-texttospeech>=2.16.0
moviepy==1.0.3
imageio-ffmpeg>=0.4.9
pydub>=0.25.1
pillow>=10.0.0
numpy>=1.26.0
google-cloud-firestore>=2.15.0
